Optmize route with car autonomy and charging stations

+++ src/jsMain/kotlin/com/github/enteraname74/project/model/Car.kt
@@ -10,5 +10,5 @@ data class Car(
     val make: String,
     val model: String,
     val version: String,
-    val autonomy: Int
+    val autonomy: Float
 )
